[Refactoring]Rename to kernelboot
Conflicts with `uefi::table::boot`.

Fixes: #143
